fix(links): permissive combine strategy
This changes the way links.combine works in the following ways:
- if `tos` is empty, the result is a list of links without To;
- if `froms` is empty, the result is a list of links without From;
- don't inherit the `tos` from the `froms` if missing
- otherwise, zip the two

This means that now, the `Default` values take precedence over the
`from` for a link. This might not be ideal, but can be fixed later.

It also means, that `Defaults` without a `From` (e.g. only setting the
`To`) will work.

A better solution might be to have `defaults` and `overrides` as global
settings. `defaults` being applied _before_ the combination,
`overrides`, after.